Renamed tags no longer include errant comma
Turns out when you have a custom tag, the tag will be printed out with
an extra comma to join it to the arguments. For example:

para "Hello, world!", class: "main"

tag "custom", "Hello, world!", class: "main"

p (and select) were not being included in the list of non-custom tags.
They were listed under RENAMED_TAGS inside Lucky. Including their tag
values in the list of tags was the fix.

Co-authored-by: Eliza Mitchell <eliza.j.mitchell@gmail.com>